Pass Fixture tests
closes #504
